+++ archive/notifier/src/main/scala/uk/ac/wellcome/platform/archive/notifier/models/NotifierConfig.scala
@@ -1,0 +1,19 @@
+package uk.ac.wellcome.platform.archive.notifier.models
+
+import uk.ac.wellcome.messaging.sns.SNSConfig
+import uk.ac.wellcome.messaging.sqs.SQSConfig
+import uk.ac.wellcome.monitoring.MetricsConfig
+import uk.ac.wellcome.platform.archive.common.modules.{
+  CloudwatchClientConfig,
+  SQSClientConfig,
+  SnsClientConfig
+}
+
+case class NotifierConfig(
+  cloudwatchClientConfig: CloudwatchClientConfig,
+  sqsClientConfig: SQSClientConfig,
+  sqsConfig: SQSConfig,
+  snsClientConfig: SnsClientConfig,
+  snsConfig: SNSConfig,
+  metricsConfig: MetricsConfig
+)
