+++ archive/notifier/src/test/scala/uk/ac/wellcome/platform/archive/notifier/fixtures/LocalWireMockFixture.scala
@@ -1,0 +1,13 @@
+package uk.ac.wellcome.platform.archive.notifier.fixtures
+
+import com.github.tomakehurst.wiremock.client.WireMock
+import uk.ac.wellcome.test.fixtures.TestWith
+
+trait LocalWireMockFixture {
+  def withLocalWireMockClient[R](host: String = "localhost", port: Int = 8080)(
+    testWith: TestWith[WireMock, R]): R = {
+    val wireMock = new WireMock(host, port)
+    wireMock.resetRequests()
+    testWith(wireMock)
+  }
+}
