+++ build.sbt
@@ -186,7 +186,7 @@ lazy val archive_common = doServiceSetup(project, "archive/common")
 lazy val archivist = doServiceSetup(project, "archive/archivist")
   .dependsOn(archive_common % "compile->compile;test->test")
 
-lazy val call_bäckerei = doServiceSetup(project, "archive/call_bäckerei")
+lazy val notifier = doServiceSetup(project, "archive/notifier")
   .dependsOn(archive_common % "compile->compile;test->test")
   .settings(libraryDependencies ++= Dependencies.wiremockDependencies)
 
@@ -236,7 +236,7 @@ lazy val root = (project in file("."))
 
     archive_common,
     archivist,
-    call_bäckerei,
+    notifier,
     registrar,
     progress
   )
